Feature: add authorization to family & person controllers

+++ FamilyTree_Backend/API/Controllers/PersonController.cs
@@ -18,19 +18,24 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using FamilyTreeBackend.Presentation.API.Handlers;
 
 namespace FamilyTreeBackend.Presentation.API.Controllers
 {
     [Area("person-management")]
-    //Commented for testing
-    //[Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
+    [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
     public class PersonController : BaseController
     {
         private readonly IPersonService _personService;
+        private readonly ITreeAuthorizationService _authorizationService;
 
-        public PersonController(UserManager<ApplicationUser> userManager, IPersonService personService) : base(userManager)
+        public PersonController(
+            UserManager<ApplicationUser> userManager,
+            IPersonService personService,
+            ITreeAuthorizationService authorizationService) : base(userManager)
         {
             _personService = personService;
+            _authorizationService = authorizationService;
         }
 
         [HttpPost("person/{personId}/parent")]
@@ -38,6 +43,18 @@ namespace FamilyTreeBackend.Presentation.API.Controllers
         [SwaggerResponse(200, Type = typeof(HttpResponse<AddNewParentToPersonResponseModel>), Description = "Return family with the new parent inside")]
         public async Task<IActionResult> AddNewParent(long personId, [FromBody] PersonInputModel input)
         {
+            var authorizationResult = await _authorizationService.AuthorizeWithPersonAsync(
+                User,
+                personId,
+                PersonOperations.Create
+                );
+            if (!authorizationResult.Succeeded)
+            {
+                return Unauthorized(new HttpResponse<AuthorizationFailure>(
+                    authorizationResult.Failure,
+                    GenericResponseStrings.Person_NoPermissionEdit));
+            }
+
             // Carry on with the business logic
             var model = new AddNewParentToPersonModel()
             {
@@ -55,7 +72,18 @@ namespace FamilyTreeBackend.Presentation.API.Controllers
         [SwaggerResponse(200, Type = typeof(HttpResponse<PersonDTO>), Description = "Returns the new spouse")]
         public async Task<IActionResult> AddNewSpouse(long personId, [FromBody] PersonInputModel input)
         {
-           
+            var authorizationResult = await _authorizationService.AuthorizeWithPersonAsync(
+                User,
+                personId,
+                PersonOperations.Create
+                );
+            if (!authorizationResult.Succeeded)
+            {
+                return Unauthorized(new HttpResponse<AuthorizationFailure>(
+                    authorizationResult.Failure,
+                    GenericResponseStrings.Person_NoPermissionEdit));
+            }
+
             // Carry on with the business logic
             var model = new AddNewSpouseToPersonModel()
             {
@@ -63,7 +91,6 @@ namespace FamilyTreeBackend.Presentation.API.Controllers
                 SpouseInfo = input
             };
             PersonDTO result = await _personService.AddNewSpouse(model);
-
             return Ok(new HttpResponse<PersonDTO>(result, GenericResponseStrings.PersonController_AddSpouseToPersonSuccessful));
            
         }
@@ -73,8 +100,27 @@ namespace FamilyTreeBackend.Presentation.API.Controllers
         [SwaggerResponse(200, Type = typeof(HttpResponse<AddNewChildToFamilyResponseModel>), Description = "Return the info of the new child, along with the new parent created on the spot (if have)")]
         public async Task<IActionResult> AddNewChild([FromBody] AddNewChildToFamilyModel input)
         {
-            var result = await _personService.AddNewChild(input);
+            var authorizationResult = await _authorizationService.AuthorizeWithPersonAsync(
+                User,
+                input.FatherId.Value,
+                PersonOperations.Create
+                );
+            if (!authorizationResult.Succeeded)
+            {
+                authorizationResult = await _authorizationService.AuthorizeWithPersonAsync(
+                HttpContext.User,
+                input.MotherId.Value,
+                PersonOperations.Create
+                );
+            }
+            if (!authorizationResult.Succeeded)
+            {
+                return Unauthorized(new HttpResponse<AuthorizationFailure>(
+                   authorizationResult.Failure,
+                   GenericResponseStrings.Person_NoPermissionEdit));
+            }
 
+            var result = await _personService.AddNewChild(input);
             return Ok(new HttpResponse<AddNewChildToFamilyResponseModel>(result, GenericResponseStrings.PersonController_AddChildToPersonSuccessful));
         }
 
@@ -83,6 +129,18 @@ namespace FamilyTreeBackend.Presentation.API.Controllers
             Description = "Return the info of the person")]
         public async Task<IActionResult> FindPerson(long personId)
         {
+            var authorizationResult = await _authorizationService.AuthorizeWithPersonAsync(
+                User,
+                personId,
+                PersonOperations.Read
+                );
+            if (!authorizationResult.Succeeded)
+            {
+                return Unauthorized(new HttpResponse<AuthorizationFailure>(
+                    authorizationResult.Failure,
+                    GenericResponseStrings.Person_NoPermissionRead));
+            }
+
             PersonModel personModel = await _personService.GetPerson(personId);
             return Ok(new HttpResponse<PersonModel>(
                 personModel, GenericResponseStrings.PersonController_FindPersonSuccessful));
@@ -93,6 +151,18 @@ namespace FamilyTreeBackend.Presentation.API.Controllers
             Description = "Return the children of the person")]
         public async Task<IActionResult> FindChildren(long personId)
         {
+            var authorizationResult = await _authorizationService.AuthorizeWithPersonAsync(
+                User,
+                personId,
+                PersonOperations.Create
+                );
+            if (!authorizationResult.Succeeded)
+            {
+                return Unauthorized(new HttpResponse<AuthorizationFailure>(
+                    authorizationResult.Failure,
+                    GenericResponseStrings.Person_NoPermissionRead));
+            }
+
             IEnumerable<PersonModel> result = await _personService.GetPersonChildren(personId);
             return Ok(new HttpResponse<IEnumerable<PersonModel>>(result, $"List of children belonging to person: {personId}"));
         }
@@ -102,6 +172,18 @@ namespace FamilyTreeBackend.Presentation.API.Controllers
             Description = "Remove the person with given Id")]
         public async Task<IActionResult> RemovePerson(long personId)
         {
+            var authorizationResult = await _authorizationService.AuthorizeWithPersonAsync(
+                User,
+                personId,
+                PersonOperations.Create
+                );
+            if (!authorizationResult.Succeeded)
+            {
+                return Unauthorized(new HttpResponse<AuthorizationFailure>(
+                    authorizationResult.Failure,
+                    GenericResponseStrings.Person_NoPermissionDelete));
+            }
+
             await _personService.RemovePerson(personId);
             return Ok(GenericResponseStrings.PersonController_RemovePersonSuccessful);
         }
@@ -111,6 +193,17 @@ namespace FamilyTreeBackend.Presentation.API.Controllers
             Description = "Update the info of the person")]
         public async Task<IActionResult> UpdatePersonInfo(long personId, [FromBody] PersonInputModel updatedPersonModel)
         {
+            var authorizationResult = await _authorizationService.AuthorizeWithPersonAsync(
+                User,
+                personId,
+                PersonOperations.Create
+                );
+            if (!authorizationResult.Succeeded)
+            {
+                return Unauthorized(new HttpResponse<AuthorizationFailure>(
+                    authorizationResult.Failure,
+                    GenericResponseStrings.Person_NoPermissionEdit));
+            }
 
             PersonModel personModel = await _personService.UpdatePersonInfo(personId, updatedPersonModel);
             var response = new HttpResponse<PersonModel>(personModel, GenericResponseStrings.PersonController_UpdatePersonSuccessful);
